+++ genrl/deep/agents/a2c/__init__.py
@@ -1,1 +1,1 @@
-from genrl.deep.agents.a2c.a2c_temp import A2C  # noqa
+from genrl.deep.agents.a2c.a2c import A2C  # noqa
