+++ genrl/deep/agents/vpg/__init__.py
@@ -1,1 +1,1 @@
-from genrl.deep.agents.vpg.vpg_temp import VPG  # noqa
+from genrl.deep.agents.vpg.vpg import VPG  # noqa
