+++ genrl/deep/common/base_class.py
@@ -1,5 +1,5 @@
 from abc import ABC
-from typing import Any, Dict, Optional, Tuple, Union
+from typing import Any, Dict, Tuple
 
 import numpy as np
 import torch
